add Authenticate method to serve for pre-forwarding authn
Upcoming work to instrument the rate of RPC requests by consumer (and eventually
rate limit) require that we authenticate a RPC request before forwarding. Add a
new top-level Authenticate method to the server and have it return an
AuthenticatedIdentity struct. RPC handlers will use the relevant fields of this
identity for performing authorization.